Check STATICFILES_DIRS and collect NPM-built assets if necessary
The 'deploystatic' command will now check if a given modified file is in one of the configured STATICFILES_DIRS (as opposed to just checking the app-specific 'static' dirs. It will also check if any file in NPM_ROOT_PATH was modified, and if so will collect the distribution versions of these assets (assuming they've already been built via the 'npm' command).